BZ1098463 ISPN-4293 PR for ISPN-4131 also caused a perf drop in tx mode
To fix ISPN-4131 we have to keep track of the highest tx id removed from
the completed txs map by the cleanup thread. The initial fix thus required
accessing two maps both to check if a tx was already completed and to mark
a tx as completed.

I was able to eliminate the second map lookup for marking the tx as
completed in the commit command, reducing the overhead somewhat.