BZ1108451 Takes a long time to start up all clusters which has a lot of segments
This is mainly caused by the myriad of internal copies performed by the CopyOnWriteArraySet used in InboundTransferTask for tracking the segments being requested during ST. This collection is updated way too frequently if we have a large number of segments. It seems that using a simple HashMap and doing external synchronization works better both in terms of time and memory.