BZ1148748 Removed RemoteCacheStoreIT#testSocketTimeoutForRemoteStore
The test was unreliable, because the low socket timeout was also used
on cache loader start - the RemoteStore has it's own RemoteCacheManager
used to communicate with the remote server, and on cache loader startup it
tries to ping the remote server to verify it's ok, but this can often fail
because the socket timeout is low. The test then fails because the cache
isn't even started.

Conflicts:
	server/integration/testsuite/src/test/resources/config/infinispan/rcs-local.xml